refactor(workflow): remove redundant functions and improve job display
This commit refactors the timesheet entry grid and job section to improve the display of job information and remove redundant code.  The changes improve code readability and maintainability.  The job status icons are now consistently handled, and the job list is updated to use an accordion for better organization.  Unnecessary validation and tracking functions are removed, simplifying the codebase.
