+++ workflow/templates/time_entries/timesheet_entry.html
@@ -92,11 +92,12 @@
     {{ staff_member_json|safe }}
     </script>
 
-
-    <!-- Current jobs section -->
-    <div id="current-jobs" class="container mt-4 mb-4">
-        <h4>Current Jobs</h4>
-        <div id="jobs-list" class="list-group">
+    <div id="current-jobs" class="container mb-4 border border-dark rounded">   
+        <h4 class="mt-3">Current Jobs</h4>
+        <div class="accordion" id="jobsAccordion">
+            <div id="jobs-list" class="accordion-body">
+                <!-- Dynamic job items will be loaded here -->
+            </div>
         </div>
     </div>
